Vacuum suction attack started
Added suction attack, not finished yet. A cone shaped hitbox rotates around the boss's head and sucks the player towards it. Currently, it needs a sprite or visible shape, and it's always active. It will be added to the boss's states as an additional attack when the player is out of range. Only affects velocity on the frame the player enters the hitbox - need to fix this.